+++ modules/processing_helpers.py
@@ -97,10 +97,10 @@ def get_sampler_name(sampler_index: int, img: bool = False) -> str:
     if len(sd_samplers.samplers) > sampler_index:
         sampler_name = sd_samplers.samplers[sampler_index].name
     else:
-        sampler_name = "UniPC"
+        sampler_name = "Default"
         shared.log.warning(f'Sampler not found: index={sampler_index} available={[s.name for s in sd_samplers.samplers]} fallback={sampler_name}')
     if img and sampler_name == "PLMS":
-        sampler_name = "UniPC"
+        sampler_name = "Default"
         shared.log.warning(f'Sampler not compatible: name=PLMS fallback={sampler_name}')
     return sampler_name